refactor: consolidate frontend dedup + type/quality mediums (batch 6)
Low-risk maintainability cleanups from the review (behavior preserved):
- Delete local formatHexValue in block-detail-client; reuse
  hexToNumber + formatNumberWithCommas from helpers.
- Single useIsMobile hook in app/lib/hooks.ts; transaction-view and the
  address panels share it (drops an inline resize listener).
- Single EpochInfo type in app/types/epoch.ts; home-client,
  validators-client, and EpochInfoPanel import it (was duplicated 3x).
- Single tx-hash matcher (TX_HASH_RE / isTxHash in helpers); tx page,
  pending-tx page, and searchResolver reuse it (was 4 copies).
- Remove unnecessary 'use client' from Footer and address/layout
  (pure-display, no client-only APIs).
- Type stray any: BlockResult.transactions unknown[], richlist
  RichlistEntry[], pending-tx catch uses axios.isAxiosError narrowing.

Deliberately NOT consolidated: the three local formatTimestamp variants
render genuinely different formats (12h AM/PM vs 24h vs toUTCString), so
merging them would change displayed output. Left as-is.

Net: lint warnings 23 -> 18, 0 errors; tsc clean; 46 unit tests pass.
