+++ ExplorerFrontend/app/address/layout.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import type { ReactNode } from "react";
 
 export default function AddressLayout({
