+++ ExplorerFrontend/app/components/Footer.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import Link from 'next/link';
 
 export default function Footer(): JSX.Element {
